+++ internal/helm/controller/reconcile.go
@@ -34,9 +34,9 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
-	"github.com/operator-framework/operator-sdk/pkg/helm/internal/diff"
-	"github.com/operator-framework/operator-sdk/pkg/helm/internal/types"
-	"github.com/operator-framework/operator-sdk/pkg/helm/release"
+	"github.com/operator-framework/operator-sdk/internal/helm/internal/diff"
+	"github.com/operator-framework/operator-sdk/internal/helm/release"
+	"github.com/operator-framework/operator-sdk/internal/helm/types"
 )
 
 // blank assignment to verify that HelmOperatorReconciler implements reconcile.Reconciler
