+++ internal/helm/release/manager_factory.go
@@ -28,8 +28,8 @@ import (
 	v1 "k8s.io/client-go/kubernetes/typed/core/v1"
 	crmanager "sigs.k8s.io/controller-runtime/pkg/manager"
 
-	"github.com/operator-framework/operator-sdk/pkg/helm/client"
-	"github.com/operator-framework/operator-sdk/pkg/helm/internal/types"
+	"github.com/operator-framework/operator-sdk/internal/helm/client"
+	"github.com/operator-framework/operator-sdk/internal/helm/types"
 )
 
 // ManagerFactory creates Managers that are specific to custom resources. It is
